Extract NexusTestSupport to nexus-test-common module
