+++ nexus/nexus-core-plugins/nexus-ldap-plugin-parent/nexus-ldap-realm-plugin/pom.xml
@@ -235,23 +235,4 @@
 
   </build>
 
-  <profiles>
-    <profile>
-      <id>with-tests</id>
-      <activation>
-        <property>
-          <name>!quick</name>
-        </property>
-      </activation>
-      <dependencies>
-        <dependency>
-          <groupId>org.sonatype.nexus</groupId>
-          <artifactId>nexus-configuration</artifactId>
-          <classifier>tests</classifier>
-        </dependency>
-      </dependencies>
-    </profile>
-  </profiles>
-
-
 </project>
